refactor(randx): move to ./internal
